refactor (user): modify user_type to ENUM.
-This commit refactors the 'user_type' column in the 'users' table and its corresponding Java entity, changing its data type from VARCHAR to ENUM. The ENUM now includes possible values 'ADMINISTRATOR' and 'USER'.
