+++ src/main/java/com/alumnione/ecommerce/dto/OrderCreationDto.java
@@ -1,19 +1,10 @@
 package com.alumnione.ecommerce.dto;
 
-import com.alumnione.ecommerce.entity.Cellphone;
 import com.alumnione.ecommerce.entity.Invoice;
-import com.alumnione.ecommerce.entity.User;
-import com.alumnione.ecommerce.entity.enums.EOrderStatus;
-import jakarta.persistence.*;
+import com.alumnione.ecommerce.enumeration.EOrderStatus;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.List;
 
 
 public record OrderCreationDto(
