+++ src/main/java/com/alumnione/ecommerce/enumeration/UserType.java
@@ -1,0 +1,6 @@
+package com.alumnione.ecommerce.enumeration;
+
+public enum UserType {
+    ADMINISTRATOR,
+    USER
+}
