+++ src/main/resources/db/migration/V16__alter_table_users_modify_userType.sql
@@ -1,0 +1,2 @@
+ALTER TABLE users
+    MODIFY COLUMN user_type ENUM('ADMINISTRATOR', 'USER') NOT NULL;
